GSdx: started a hardware independent device, will be useful for GDI/SDL/etc output later.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@4328 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/GSdx/GSDialog.cpp
@@ -129,7 +129,7 @@ void GSDialog::ComboBoxInit(UINT id, const vector<GSSetting>& settings, uint32 s
 
 	SendMessage(hWnd, CB_RESETCONTENT, 0, 0);
 
-	for(int i = 0; i < settings.size(); i++)
+	for(size_t i = 0; i < settings.size(); i++)
 	{
 		const GSSetting& s = settings[i];
